refactor: optimized movie search by id and get similar movies
